+++ src/app/contribution/contribution.module.ts
@@ -1,3 +1,4 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -33,6 +34,7 @@ import { StableProjectCreationModal } from './stable/stable-project-creation-mod
     ],
     imports: [
         CommonModule,
+        DragDropModule,
         FormsModule,
         TranslateModule,
         WidgetModule,