+++ src/app/graph/graph.module.ts
@@ -1,3 +1,4 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -53,6 +54,7 @@ import { ResourceDetailsPanel } from './widget/resource-details-panel';
     ],
     imports: [
         CommonModule,
+        DragDropModule,
         FormsModule,
         NgbDropdownModule,
         PreferencesModule,