+++ src/app/modal-dialogs/modals.module.ts
@@ -1,3 +1,4 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -46,6 +47,7 @@ import { SharedModalsServices } from './shared-modals/shared-modal.service';
     ],
     imports: [
         CommonModule, 
+        DragDropModule,
         FormsModule,
         ResourceViewModule,
         StructuresModule,