+++ src/app/search/search.module.ts
@@ -1,3 +1,4 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -14,6 +15,7 @@ import { SearchComponent } from './search.component';
     ],
     imports: [
         CommonModule,
+        DragDropModule,
         FormsModule,
         NgbDropdownModule,
         RouterModule,