+++ src/app/user/user.module.ts
@@ -1,3 +1,4 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -21,6 +22,7 @@ import { UserProfileComponent } from './user-profile.component';
 	],
 	imports: [
 		CommonModule,
+		DragDropModule,
 		FormsModule,
         ModalsModule,
 		RouterModule,