+++ ChangeLog.md
@@ -5,7 +5,7 @@ The full changelog of the platform includes also changes brought to the Semantic
 https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 1.1.2 (dd-mm-yyyy)
-  * 
+  * Made dialogs draggable
 
 # 1.1.1 (28-10-2021)
   * Added localization for French, thanks to Nathalie Vedovotto for the contribution!